+++ polaris-migrator/src/migrations/replace-sass-spacing/replace-sass-spacing.ts
@@ -1,8 +1,21 @@
 import type {FileInfo, API, Options} from 'jscodeshift';
 import postcss, {Plugin} from 'postcss';
-import valueParser, {Node, FunctionNode} from 'postcss-value-parser';
+import valueParser from 'postcss-value-parser';
 
 import {POLARIS_MIGRATOR_COMMENT} from '../../constants';
+import {
+  NamespaceOptions,
+  namespace,
+  isSassFunction,
+  hasSassFunction,
+  hasNumericOperator,
+  hasSassInterpolation,
+  removeSassInterpolation,
+  hasNegativeSassInterpolation,
+  replaceNegativeSassInterpolation,
+  createInlineComment,
+} from '../../utilities/sass';
+import {isKeyOf} from '../../utilities/type-guards';
 
 const spacingMap = {
   none: '--p-space-0',
@@ -15,31 +28,12 @@ const spacingMap = {
   'extra-loose': '--p-space-8',
 };
 
-const isSpacing = (spacing: unknown): spacing is keyof typeof spacingMap =>
-  Object.keys(spacingMap).includes(spacing as string);
-
-function isNumericOperator(node: Node): boolean {
-  return (
-    node.value === '+' ||
-    node.value === '-' ||
-    node.value === '*' ||
-    node.value === '/' ||
-    node.value === '%'
-  );
-}
-
 const processed = Symbol('processed');
 
-interface PluginOptions extends Options {
-  namespace?: string;
-}
+interface PluginOptions extends Options, NamespaceOptions {}
 
 const plugin = (options: PluginOptions = {}): Plugin => {
-  const namespace = options?.namespace || '';
-  const functionName = namespace ? `${namespace}.spacing` : 'spacing';
-  const isSpacingFn = (node: Node): node is FunctionNode => {
-    return node.type === 'function' && node.value === functionName;
-  };
+  const namespacedSpacing = namespace('spacing', options);
 
   return {
     postcssPlugin: 'ReplaceSassSpacing',
@@ -47,20 +41,27 @@ const plugin = (options: PluginOptions = {}): Plugin => {
       // @ts-expect-error - Skip if processed so we don't process it again
       if (decl[processed]) return;
 
-      const parsed = valueParser(decl.value);
+      const parsedValue = valueParser(decl.value);
 
-      let containsSpacingFn = false;
-      let containsCalculation = false;
+      // Convert -#{spacing()} to -1 * #{spacing()}
+      if (hasNegativeSassInterpolation(decl.value)) {
+        replaceNegativeSassInterpolation(parsedValue);
+      }
 
-      parsed.walk((node) => {
-        if (isSpacingFn(node)) containsSpacingFn = true;
-        if (isNumericOperator(node)) containsCalculation = true;
+      // Remove #{} from spacing()
+      if (hasSassInterpolation(parsedValue.toString())) {
+        removeSassInterpolation(namespacedSpacing, parsedValue);
+      }
 
-        if (!isSpacingFn(node)) return;
+      // Now we can check if the value is a spacing() function
+      if (!hasSassFunction(namespacedSpacing, parsedValue)) return;
+
+      parsedValue.walk((node) => {
+        if (!isSassFunction(namespacedSpacing, node)) return;
 
         const spacing = node.nodes[0]?.value ?? '';
 
-        if (!isSpacing(spacing)) return;
+        if (!isKeyOf(spacingMap, spacing)) return;
         const spacingCustomProperty = spacingMap[spacing];
 
         node.value = 'var';
@@ -71,18 +72,19 @@ const plugin = (options: PluginOptions = {}): Plugin => {
             sourceIndex: node.nodes[0]?.sourceIndex ?? 0,
             sourceEndIndex: spacingCustomProperty.length,
           },
-          ...node.nodes.slice(1),
         ];
       });
 
-      if (containsSpacingFn && containsCalculation) {
+      if (hasNumericOperator(parsedValue)) {
         // Insert comment if the declaration value contains calculations
-        decl.before(postcss.comment({text: POLARIS_MIGRATOR_COMMENT}));
         decl.before(
-          postcss.comment({text: `${decl.prop}: ${parsed.toString()};`}),
+          createInlineComment(POLARIS_MIGRATOR_COMMENT, {prose: true}),
+        );
+        decl.before(
+          createInlineComment(`${decl.prop}: ${parsedValue.toString()};`),
         );
       } else {
-        decl.value = parsed.toString();
+        decl.value = parsedValue.toString();
       }
 
       // @ts-expect-error - Mark the declaration as processed