+++ polaris-cli/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @shopify/plugin-polaris
 
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies [[`85c0c3290`](https://github.com/Shopify/polaris/commit/85c0c329003c0d234f339164c0d2940aff4f6b26), [`141746b5b`](https://github.com/Shopify/polaris/commit/141746b5b4dc2f99b0c729883e98f4210f5a16c0)]:
+  - @shopify/polaris-migrator@0.7.0
+
+## 0.0.14
+
+## 0.0.13
+
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies [[`dc2990acf`](https://github.com/Shopify/polaris/commit/dc2990acf1071c2bae352bce6fc5a28092b9f52f), [`8fb215836`](https://github.com/Shopify/polaris/commit/8fb2158368945ad1e58b6e3ff74567b80b652757), [`60191f353`](https://github.com/Shopify/polaris/commit/60191f353d1ef86f1eef167d36c0b040f21b72d6), [`6d82e8965`](https://github.com/Shopify/polaris/commit/6d82e896587bacec60471e06fa81b2c61518551f), [`0591db26a`](https://github.com/Shopify/polaris/commit/0591db26a4d13a5640245e433f61f77bf1be88d4), [`699b70713`](https://github.com/Shopify/polaris/commit/699b707132bfe486785955084961bcb91753fb2a), [`8194e174a`](https://github.com/Shopify/polaris/commit/8194e174ab45e1995f693e786d6925ab79717c1c), [`933a9d71b`](https://github.com/Shopify/polaris/commit/933a9d71b71b172657bbaf050cfc8793e39a49c0), [`a0fa96ed4`](https://github.com/Shopify/polaris/commit/a0fa96ed4eaf53652feb5ee0d79aeec1c84f5d12)]:
+  - @shopify/polaris-migrator@0.6.0
+
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [[`56c82ee8d`](https://github.com/Shopify/polaris/commit/56c82ee8d15a58ef5e68b04968b126b93aa8dec9)]:
+  - @shopify/polaris-migrator@0.5.0
+
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [[`57e5f95c1`](https://github.com/Shopify/polaris/commit/57e5f95c1e402fda3202babfdce25b131b0b9165)]:
+  - @shopify/polaris-migrator@0.4.1
+
+## 0.0.9
+
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [[`fbf2f8832`](https://github.com/Shopify/polaris/commit/fbf2f88320ad6c5767c37d18d76bfa82babf71cb), [`b31f51f25`](https://github.com/Shopify/polaris/commit/b31f51f259fefd8b119c5242e2ab1528c3a6cff8)]:
+  - @shopify/polaris-migrator@0.4.0
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [[`5a1f45f7a`](https://github.com/Shopify/polaris/commit/5a1f45f7a29cf6994394d7fab6f007f850d956ad), [`c958899c7`](https://github.com/Shopify/polaris/commit/c958899c73829bf1c1fa512946ffa3db7f06a2d0)]:
+  - @shopify/polaris-migrator@0.3.0
+
 ## 0.0.6
 
 ## 0.0.5